Scrapers/Disney/Production/Elbert_reservation/Disney/dis_requirements.txt
alembic==1.7.5
amqp==5.1.0
anyio==3.4.0
apache-airflow==2.2.5
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
apturl==0.5.2
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==20.3.0
Automat==0.8.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==5.0.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachelib==0.6.0
cachetools==4.2.4
cajarename==19.7.15
cattrs==1.5.0
celery==5.1.2
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
cloud-init==22.1
colorama==0.4.4
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
commonregex==1.5.4
configobj==5.0.6
connexion==2.13.0
constantly==15.1.0
croniter==1.0.15
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
DateTime==4.3
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
distro==1.5.0
distro-info===0.23ubuntu1
Django==1.3.7
django-indexer==0.3.0
django-kombu==0.9.4
django-paging==0.2.5
django-templatetag-sugar==1.0
dnspython==2.1.0
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
ecdsa==0.17.0
email-validator==1.1.3
entrypoints==0.3
eventlet==0.33.0
executing==0.8.3
expandvars==0.7.0
fastavro==1.4.8
fasteners==0.14.1
fastjsonschema==2.15.3
Flask==1.1.2
Flask-AppBuilder==3.4.1
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==1.0.0
fonttools==4.33.3
future==0.18.2
gcp==0.2.1
google==3.0.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gpg===1.13.1-unknown
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
html-table-parser-python3==0.2.0
html2text==2020.1.16
httpcore==0.13.7
httplib2==0.20.2
httpx==0.19.0
humanize==3.14.0
hyperlink==19.0.0
idna==2.10
imap-tools==0.53.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
incremental==16.10.1
inflection==0.5.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso8601==1.0.2
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==3.0.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
jupyterthemes==0.20.0
keyring==18.0.1
kiwisolver==1.4.2
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lesscpy==0.15.0
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Magnus==1.0.3
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mate-hud==19.10.0
mate-menu==20.4.1
mate-tweak==20.4.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.11
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
onboard==1.4.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openlineage-airflow==0.6.2
openlineage-integration-common==0.6.2
openlineage-python==0.6.2
outcome==1.1.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
paramiko==2.7.1
parsedatetime==2.6
parso==0.8.3
pathlib2==2.3.6
pbr==5.8.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
ply==3.11
prison==0.2.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.2
psutil==5.8.0
psycopg2==2.9.3
ptyprocess==0.7.0
pulsemixer==1.5.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.21.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.9.0
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-daemon==2.3.0
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.14.0
python-jose==3.3.0
python-nvd3==0.15.0
python-slugify==4.0.1
python-xapp==1.8.1
python-xlib==0.23
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyVirtualDisplay==3.0
pyxdg==0.26
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
raven==6.10.0
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rfc3986==1.5.0
rich==10.16.0
rsa==4.8
SecretStorage==2.3.1
selenium==4.1.3
Send2Trash==1.8.0
sentry==2.1.2
service-identity==18.1.0
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
South==1.0.2
sql-metadata==2.0.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
sqlparse==0.4.2
ssh-import-id==5.7
stack-data==0.2.0
statsd==3.3.0
swagger-ui-bundle==0.0.9
systemd-python==234
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.13.3
testresources==2.0.1
text-unidecode==1.3
texttable==1.6.2
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
typing-extensions==4.0.1
typing-inspect==0.7.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wrapt==1.14.0
wsproto==1.1.0
WTForms==2.3.3
xkit==0.0.0
zipp==3.6.0
zope.interface==4.7.1
